clients/window: Update min_allocation for smaller widths/heights
This makes sure we update min_allocation in situations where the
width and height passed is smaller than the one set previously
(obviously except for the first time). This has the side effect of
not overwriting pending_allocations such that we correctly resize when
passed a width/height smaller the ones set-up in the past.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
(cherry picked from commit 955511809558424bb74279df0f88465087e35723)
Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>