hmi-controller: Do not call hmi_controller_destroy() from _create()
Move the destroy listener registration to the top, to avoid having
to call hmi_controller_destroy() for cleanup when it fails.

This would segfault trying to remove the destroy listener from a list
when its link was never initialized or inserted into any list.

The failure case can be hit by setting both modules=hmi-controller.so in
weston.ini and --modules=hmi-controller.so on the command line.

Signed-off-by: Philipp Zabel <p.zabel@pengutronix.de>